touchscreen: synaptics_dsx_force/nvt/fts/fts_521: add pm_qos_request in irq handler
We found there is a delay when CPU get irq and read data from touch,
so we add pm_qos_request to let CPU stay awake while doing irq handler.

francisco: Adapted for NVT and FTS panels on the fuckofone.
Hlcpereira: Import to synaptics_dsx_force and adapt for fts_521

Bug: 122431169
Test: Check if touch latency improved.

Signed-off-by: davidycchen <davidycchen@google.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>
Signed-off-by: Henrique Pereira <hlcpereira@pixelexperience.org>